+++ changelog.md
@@ -44,6 +44,7 @@
     * Added Fluorspar
     * Added Gadolinium
     * Added Hard Carbon
+    * Added Jade
     * Added Lithium Manganese Dioxide
     * Added Manganosite
     * Added Nepheline
@@ -60,9 +61,10 @@
     * Added Thallium
     * Added Thermoconducting Alloy
     * Added Tough Alloy
+    * Added Ulexite
     * Added Vibrant Alloy
     * Added Villiaumite
-    * Added Zircalloy
+    * Added Zircaloy
   * Unifications
     * Unified GT Certus Quartz and AE2 Certus Quartsz
     * Unified GT Enriched Certus Quartz with AE2 Charged Certus Quartz
@@ -75,15 +77,61 @@
     * Ostrum now requires an Hv EBF with Kanthal Coils
     * Calorite now requires an Ev EBF with Nichrome Coils
     * Moon Deepslate no longer contains quicklime, can now be processed in an electrolyzer
+    * Added ore to Antimony
+    * Added ore to Arsenic
+    * Added ore to Bismuth
     * Added ore to Biotite
+    * Added ore to Borax
+    * Added ore to Chromium
+    * Added ore to Draconium
+    * Added ore to Enriched Naquadah
+    * Added ore to Mithril
+    * Added ore to Perlite
+    * Added ore to Rutile
+    * Added ore to Titanium
+    * Added ore to Uvarovite
 * Ore Veins
   * Additions
+    * Added Ardite (Ardite, Amethyst, Rutile) vein to Venus
+    * Added Arsenic (Arsenic, Bismuth, Antimony) vein to Mars and Mercury
+    * Added Bauxite (Bauxite, Ilmenite, Aluminium) vein to Mercury
+    * Added Beryllium (Beryllium, Emerald, Thorium) vein to Mars and Venus
+    * Added Borax (Borax, Ulexite) vein to the Moon
     * Added Carobbiite (Carobbiite) vein to Mars
-    * Added Nepheline (Nepheline, Biotite, Titanite, Villiaumite) vein to Mars
+    * Added Cassiterite (Tin, Cassiterite) vein to the Moon and Venus
+    * Added Copper (Chalcopyrite, Iron, Pyrite, Copper) vein to the Moon
+    * Added Diamond (Graphite, Diamond, Coal) vein to Mercury
+    * Added Draconium (Draconium, Electrotine, Jade) vein to Mercury
     * Added Fluorspar (Fluorspar, Barite, Calcite) vein to Mars
+    * Added Galena (Galena, Silver, Lead) vein to the Moon, Mars and Venus
+    * Added Gold (Magnetite, Gold, Vanadium Magnetite) vein to Mars
+    * Added Ilmenite (Ilmenite, Chromite, Uvarovite, Perlite) vein to the Moon and Mercury
+    * Added Iridium (Iridium, Palladium, Mithril) vein to Mercury and Venus
+    * Added Iron (Goethite, Yellow Limonite, Hematite, Malachite) vein to Mars and Mercury
+    * Added Molybdenum (Wulfenite, Molybdenite, Molybdenum, Powellite) vein to the Moon and Mercury
+    * Added Monazite (Bastnasite, Monazite, Neodymium) vein to the Moon and Venus
+    * Added Naquadah (Naquadah, Enriched Naquadah) vein to Mercury and Venus
+    * Added Nickel (Garnierite, Nickel, Cobaltite, Pentlandite) vein to Mars and Venus
+    * Added Palladium (Platinum, Chromium ,Sheldonite, Palladium) vein to Mercury
+    * Added Pitchblende (Pitchblende, Uraninite) to Venus
+    * Added Quartzite (Quartzite, Barite, Certus Quartz) vein to the Moon, Mars and Venus
+    * Added Redstone (Redstone, Ruby, Cinnabar) vein to Mars, Mercury and Venus
     * Added Rhodochrosite (Rhodochrosite, Silver) vein to Mars
+    * Added Rutile (Rutile, Titanium, Bauxite, Calorite) vein to Venus
+    * Added Salts (Rock Salt, Salt, Lepidolite, Spodumene) vein to Mars
+    * Added Sulfur (Sulfur, Pyrite, Sphalerite) vein to Mars and Venus
+    * Added Tetrahedrite (Tetrahedrite, Copper, Stibnite) vein to Mars and Venus
+    * Added Nepheline (Nepheline, Biotite, Titanite, Villiaumite) vein to Mars
   * Modifications
-    * Fixed Naquadah, Sheldonite, Pitchblende, Scheelite veins not spawning on Mars
+    * Fixed Naquadah, Pitchblende, Scheelite veins not spawning on Mars
+    * Moved Sheldonite vein to the Moon
+* Bedrock Fluid Deposits
+  * Added Chlorobenzene deposit to the Moon
+  * Added CO2 deposit to Venus
+  * Added Helium-3 deposit to the Moon and Mercury
+  * Added Molten Iron deposit to Mercury
+  * Added Molten Lead deposit to Venus
+  * Added Sulfuric Acid deposit to Venus
 * Resource Packs
   * Additions
     * Added Better GUI for Sophisticated Backpacks